Fix bar animation. Document some props.
