+++ packages/x-charts/src/BarChart/BarLabel/BarLabel.tsx
@@ -2,9 +2,7 @@
 import * as React from 'react';
 import { styled, useThemeProps } from '@mui/material/styles';
 import PropTypes from 'prop-types';
-import { interpolateNumber } from '@mui/x-charts-vendor/d3-interpolate';
-import { useIsHydrated } from '../../hooks/useIsHydrated';
-import { useAnimate } from '../../internals/animation/useAnimate';
+import { useAnimateBarLabel } from '../../hooks/animation/useAnimateBarLabel';
 import { barLabelClasses } from './barLabelClasses';
 import { BarLabelOwnerState } from './BarLabel.types';
 
@@ -30,76 +28,15 @@ export const BarLabelComponent = styled('text', {
   },
 }));
 
-type UseAnimateBarLabelParams = Pick<
-  BarLabelProps,
-  'x' | 'y' | 'width' | 'height' | 'layout' | 'skipAnimation'
-> & {
-  ref?: React.Ref<SVGTextElement>;
-};
-type UseAnimateBarLabelReturn = {
-  ref: React.Ref<SVGTextElement>;
-} & Pick<BarLabelProps, 'x' | 'y' | 'width' | 'height'>;
-type BarLabelInterpolatedProps = Pick<UseAnimateBarLabelParams, 'x' | 'y' | 'width' | 'height'>;
-
-function barLabelPropsInterpolator(from: BarLabelInterpolatedProps, to: BarLabelInterpolatedProps) {
-  const interpolateX = interpolateNumber(from.x, to.x);
-  const interpolateY = interpolateNumber(from.y, to.y);
-  const interpolateWidth = interpolateNumber(from.width, to.width);
-  const interpolateHeight = interpolateNumber(from.height, to.height);
-
-  return (t: number) => {
-    return {
-      x: interpolateX(t),
-      y: interpolateY(t),
-      width: interpolateWidth(t),
-      height: interpolateHeight(t),
-    };
-  };
-}
-
-export function useAnimateBarLabel(props: UseAnimateBarLabelParams): UseAnimateBarLabelReturn {
-  const isHydrated = useIsHydrated();
-  const initialProps = {
-    x: props.layout === 'vertical' ? props.x + props.width / 2 : props.x,
-    y: props.layout === 'vertical' ? props.y + props.height : props.y + props.height / 2,
-    width: props.width,
-    height: props.height,
-  };
-  const currentProps = {
-    x: props.x + props.width / 2,
-    y: props.layout === 'vertical' ? props.y + props.height / 2 : props.y + props.height / 2,
-    width: props.width,
-    height: props.height,
-  };
-
-  const ref = useAnimate(currentProps, {
-    createInterpolator: barLabelPropsInterpolator,
-    applyProps(element, animatedProps) {
-      element.setAttribute('x', animatedProps.x.toString());
-      element.setAttribute('y', animatedProps.y.toString());
-      element.setAttribute('width', animatedProps.width.toString());
-      element.setAttribute('height', animatedProps.height.toString());
-    },
-    initialProps,
-    skip: props.skipAnimation,
-  });
-
-  const usedProps = props.skipAnimation || !isHydrated ? currentProps : initialProps;
-
-  return {
-    ref,
-    x: usedProps.x,
-    y: usedProps.y,
-    width: usedProps.width,
-    height: usedProps.height,
-  };
-}
-
 export type BarLabelProps = Omit<
   React.SVGProps<SVGTextElement>,
   'ref' | 'id' | 'x' | 'y' | 'width' | 'height'
 > &
   BarLabelOwnerState & {
+    /** x-coordinate of the stack this bar label belongs to. */
+    xOrigin: number;
+    /** y-coordinate of the stack this bar label belongs to. */
+    yOrigin: number;
     x: number;
     y: number;
     width: number;
@@ -118,6 +55,8 @@ function BarLabel(inProps: BarLabelProps) {
     classes,
     skipAnimation,
     layout,
+    xOrigin,
+    yOrigin,
     ...otherProps
   } = props;
